backend/requirements.txt
# FastAPI Backend Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# AI & ML
google-generativeai==0.3.2
openai-whisper==20231117
torch==2.4.0
transformers==4.36.0
numpy==1.24.3
librosa==0.10.1

# Audio Processing
soundfile==0.12.1
pydub==0.25.1

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Arabic NLP
arabic-reshaper==3.0.0
python-bidi==0.4.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dateutil==2.8.2
pytz==2023.3
httpx==0.25.2
websockets==12.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
